+++ plugins/CircuitExplorer/io/bbploader/reports/CompartmentData.h
@@ -30,9 +30,9 @@ class CompartmentData : public IReportData
 public:
     CompartmentData(std::unique_ptr<brion::CompartmentReport> report);
 
-    float getStartTime() const noexcept override;
-    float getEndTime() const noexcept override;
-    float getTimeStep() const noexcept override;
+    double getStartTime() const noexcept override;
+    double getEndTime() const noexcept override;
+    double getTimeStep() const noexcept override;
     std::string getTimeUnit() const noexcept override;
     std::vector<float> getFrame(double timestamp) const override;
